move ipcluster create|list to `ipython profile create|list`
* give ProfileDir its own file (core.profiledir)
* add core.profileapp for new subcommand

parallel docs updated to match
