+++ IPython/parallel/apps/iploggerapp.py
@@ -25,7 +25,7 @@ import sys
 
 import zmq
 
-from IPython.core.application import ProfileDir
+from IPython.core.profiledir import ProfileDir
 from IPython.utils.traitlets import Bool, Dict, Unicode
 
 from IPython.parallel.apps.baseapp import (
@@ -47,7 +47,7 @@ IPython controllers and engines (and your own processes) can broadcast log messa
 by registering a `zmq.log.handlers.PUBHandler` with the `logging` module. The
 logger can be configured using command line options or using a cluster
 directory. Cluster directories contain config, log and security files and are
-usually located in your ipython directory and named as "cluster_<profile>".
+usually located in your ipython directory and named as "profile_name".
 See the `profile` and `profile_dir` options for details.
 """
